added user authorization and added file upload feature

+++ app/Http/Controllers/PostsController.php
@@ -3,11 +3,24 @@
 namespace App\Http\Controllers;
 
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Storage;
 use App\Post;
 use DB;
 
 class PostsController extends Controller
 {
+
+    /**
+     * Create a new controller instance.
+     *
+     * @return void
+     */
+    public function __construct()
+    {
+        $this->middleware('auth', ['except' => ['index', 'show']]);
+    }       
+
+
     /**
      * Display a listing of the resource.
      *
@@ -46,14 +59,32 @@ class PostsController extends Controller
     {
         $this->validate($request, [
             'title' => 'required',
-            'body' => 'required'
+            'body' => 'required',
+            'cover_image' => 'image|nullable|max:1999'
         ]);
+
+        // handle file upload
+        if($request->hasFile('cover_image')) {
+            // get filename with the extension
+            $filenameWithExt = $request->file('cover_image')->getClientOriginalName();
+            // get just filename
+            $filename = pathinfo($filenameWithExt, PATHINFO_FILENAME);
+            // get just extension
+            $extension = $request->file('cover_image')->getClientOriginalExtension();
+            // filename to store
+            $fileNameToStore = $filename.'_'.time().'.'.$extension;
+            // upload image
+            $path = $request->file('cover_image')->storeAs('public/cover_images', $fileNameToStore);
+        } else {
+            $fileNameToStore = 'noimage.jpg';
+        }
 
        // Create Post
        $post = new Post;
        $post->title = $request->input('title');
        $post->body = $request->input('body');
        $post->user_id = auth()->user()->id;
+       $post->cover_image = $fileNameToStore;
        $post->save();
 
        return redirect('/posts')->with('success', 'Post Created');
@@ -81,6 +112,12 @@ class PostsController extends Controller
     public function edit($id)
     {
         $post = Post::find($id);
+
+        // check for correct user
+        if(auth()->user()->id !==$post->user_id) {
+            return redirect('/posts')->with('error', 'You are not authorized to edit this post');
+        }
+
         return view('posts.edit')->with('post', $post);
     }
 
@@ -98,10 +135,27 @@ class PostsController extends Controller
             'body' => 'required'
         ]);
 
+        // handle file upload
+        if($request->hasFile('cover_image')) {
+            // get filename with the extension
+            $filenameWithExt = $request->file('cover_image')->getClientOriginalName();
+            // get just filename
+            $filename = pathinfo($filenameWithExt, PATHINFO_FILENAME);
+            // get just extension
+            $extension = $request->file('cover_image')->getClientOriginalExtension();
+            // filename to store
+            $fileNameToStore = $filename.'_'.time().'.'.$extension;
+            // upload image
+            $path = $request->file('cover_image')->storeAs('public/cover_images', $fileNameToStore);
+        }
+
        // Update Post
        $post = Post::find($id);
        $post->title = $request->input('title');
        $post->body = $request->input('body');
+       if($request->hasFile('cover_image')) {
+           $post->cover_image = $fileNameToStore;
+       }
        $post->save();
 
        return redirect('/posts')->with('success', 'Post Updated');
@@ -116,6 +170,17 @@ class PostsController extends Controller
     public function destroy($id)
     {
         $post = Post::find($id);
+
+        // check for correct user 
+        if(auth()->user()->id !==$post->user_id) {
+            return redirect('/posts')->with('error', 'You are not authorized to delete this post');
+        }
+
+        if($post->cover_image != 'noimage.jpg') {
+            // delete image
+            Storage::delete('public/cover_images/'.$post->cover_image);
+        }
+
         $post->delete();
         return redirect('/posts')->with('success', 'Post Deleted!');
     }
